fix: move device-specific schemas into their own files
