+++ src/braket/device_schema/dwave/dwave_device_parameters_v1.py
@@ -14,8 +14,10 @@ from typing import Optional, Union
 
 from pydantic import Field
 
-from braket.device_schema.dwave.dwave_device_level_parameters_v1 import (
+from braket.device_schema.dwave.dwave_2000Q_device_level_parameters_v1 import (
     Dwave2000QDeviceLevelParameters,
+)
+from braket.device_schema.dwave.dwave_advantage_device_level_parameters_v1 import (
     DwaveAdvantageDeviceLevelParameters,
 )
 from braket.device_schema.dwave.dwave_provider_level_parameters_v1 import (
@@ -56,33 +58,3 @@ class DwaveDeviceParameters(BraketSchemaBase):
     deviceLevelParameters: Optional[
         Union[DwaveAdvantageDeviceLevelParameters, Dwave2000QDeviceLevelParameters]
     ]
-
-
-class DwaveAdvantageDeviceParameters(BraketSchemaBase):
-    """
-    This is the description of the D-Wave parameters
-
-    Attributes:
-        providerLevelParameters: Parameters that are specific to D-Wave device.
-    """
-
-    _PROGRAM_HEADER = BraketSchemaHeader(
-        name="braket.device_schema.dwave.dwave_advantage_device_parameters", version="1"
-    )
-    braketSchemaHeader: BraketSchemaHeader = Field(default=_PROGRAM_HEADER, const=_PROGRAM_HEADER)
-    deviceLevelParameters: DwaveAdvantageDeviceLevelParameters
-
-
-class Dwave2000QDeviceParameters(BraketSchemaBase):
-    """
-    This is the description of the D-Wave parameters
-
-    Attributes:
-        providerLevelParameters: Parameters that are specific to D-Wave device.
-    """
-
-    _PROGRAM_HEADER = BraketSchemaHeader(
-        name="braket.device_schema.dwave.dwave_2000Q_device_parameters", version="1"
-    )
-    braketSchemaHeader: BraketSchemaHeader = Field(default=_PROGRAM_HEADER, const=_PROGRAM_HEADER)
-    deviceLevelParameters: Dwave2000QDeviceLevelParameters
